karyon_net

Module stream

Re-exports

pub use websocket::ReadWsStream;

pub use websocket::WriteWsStream;

pub use websocket::WsStream;

Modules

websocket 🔒

Structs

ReadStream

WriteStream

Constants

BUFFER_CHUNK_SIZE 🔒

Maximum number of bytes to read at a time to construct the stream buffer.

DEFAULT_MAX_BUFFER_SIZE 🔒

The stream buffer size is limited by this value. This can be overridden by setting the KARYON_MAX_BUFFER_SIZE environment variable at compile time.

Functions

get_max_buffer_size 🔒