Crate karyon_net

Re-exports

pub use karyon_async_rustls as async_rustls;

Modules

codec

connection 🔒

endpoint 🔒

error 🔒

listener 🔒

stream 🔒

tcp

tls

transports 🔒

udp

unix

ws

Enums

Addr

Addr defines a type for an address, either IP or domain.

Endpoint

Endpoint defines generic network endpoints for karyon.

Error

Represents karyon’s Net Error

Traits

ConnListener

ConnListener is a generic network listener interface for [tcp::TcpConn], [tls::TlsConn], [ws::WsConn], and [unix::UnixConn].

Connection

Connection is a generic network connection interface for [udp::UdpConn], [tcp::TcpConn], [tls::TlsConn], [ws::WsConn], and [unix::UnixConn].

ToConn

A trait for objects which can be converted to Conn.

ToEndpoint

ToListener

A trait for objects which can be converted to Listener.

Type Aliases

Conn

Alias for Box<dyn Connection>

Listener

Alias for Box<dyn ConnListener>

Port

Port defined as a u16.

Result

Represents karyon’s Net Result